refactor(ecma/lexer): split lexer of parser
