+++ crates/swc_ecma_lexer/benches/lexer.rs
@@ -1,8 +1,8 @@
 extern crate swc_malloc;
 
 use codspeed_criterion_compat::{black_box, criterion_group, criterion_main, Bencher, Criterion};
-use swc_common::FileName;
-use swc_ecma_parser::{lexer::Lexer, StringInput, Syntax, TsSyntax};
+use swc_common::{input::StringInput, FileName};
+use swc_ecma_lexer::{lexer::Lexer, Syntax, TsSyntax};
 
 fn bench_module(b: &mut Bencher, syntax: Syntax, src: &'static str) {
     let _ = ::testing::run_test(false, |cm, _| {
@@ -21,14 +21,18 @@ fn bench_module(b: &mut Bencher, syntax: Syntax, src: &'static str) {
 fn bench_files(c: &mut Criterion) {
     c.bench_function("es/lexer/colors", |b| {
         // Copied from ratel-rust
-        bench_module(b, Default::default(), include_str!("../colors.js"))
+        bench_module(
+            b,
+            Default::default(),
+            include_str!("../../swc_ecma_parser/colors.js"),
+        )
     });
 
     c.bench_function("es/lexer/angular", |b| {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/angular-1.2.5.js"),
+            include_str!("../../swc_ecma_parser/benches/files/angular-1.2.5.js"),
         )
     });
 
@@ -36,7 +40,7 @@ fn bench_files(c: &mut Criterion) {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/backbone-1.1.0.js"),
+            include_str!("../../swc_ecma_parser/benches/files/backbone-1.1.0.js"),
         )
     });
 
@@ -44,7 +48,7 @@ fn bench_files(c: &mut Criterion) {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/jquery-1.9.1.js"),
+            include_str!("../../swc_ecma_parser/benches/files/jquery-1.9.1.js"),
         )
     });
 
@@ -52,14 +56,14 @@ fn bench_files(c: &mut Criterion) {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/jquery.mobile-1.4.2.js"),
+            include_str!("../../swc_ecma_parser/benches/files/jquery.mobile-1.4.2.js"),
         )
     });
     c.bench_function("es/lexer/mootools", |b| {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/mootools-1.4.5.js"),
+            include_str!("../../swc_ecma_parser/benches/files/mootools-1.4.5.js"),
         )
     });
 
@@ -67,7 +71,7 @@ fn bench_files(c: &mut Criterion) {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/underscore-1.5.2.js"),
+            include_str!("../../swc_ecma_parser/benches/files/underscore-1.5.2.js"),
         )
     });
 
@@ -75,12 +79,16 @@ fn bench_files(c: &mut Criterion) {
         bench_module(
             b,
             Default::default(),
-            include_str!("./files/three-0.138.3.js"),
+            include_str!("../../swc_ecma_parser/benches/files/three-0.138.3.js"),
         )
     });
 
     c.bench_function("es/lexer/yui", |b| {
-        bench_module(b, Default::default(), include_str!("./files/yui-3.12.0.js"))
+        bench_module(
+            b,
+            Default::default(),
+            include_str!("../../swc_ecma_parser/benches/files/yui-3.12.0.js"),
+        )
     });
 
     c.bench_function("es/lexer/cal-com", |b| {
@@ -90,12 +98,16 @@ fn bench_files(c: &mut Criterion) {
                 tsx: true,
                 ..Default::default()
             }),
-            include_str!("./files/cal.com.tsx"),
+            include_str!("../../swc_ecma_parser/benches/files/cal.com.tsx"),
         )
     });
 
     c.bench_function("es/lexer/typescript", |b| {
-        bench_module(b, Default::default(), include_str!("./files/typescript.js"))
+        bench_module(
+            b,
+            Default::default(),
+            include_str!("../../swc_ecma_parser/benches/files/typescript.js"),
+        )
     });
 }
