+++ crates/swc_ecma_lexer/src/lexer/comments_buffer.rs
