+++ crates/swc_ecma_lexer/src/lexer/input.rs
