+++ crates/swc_ecma_lexer/src/lexer/number.rs
