+++ crates/swc_ecma_lexer/src/lexer/table.rs
