+++ crates/swc_ecma_lexer/src/lexer/util.rs
