+++ crates/swc_ecma_lexer/src/lexer/whitespace.rs
